Rearrange public/private headers
